security/devices

+++ src/routers/security-router.ts
@@ -26,12 +26,12 @@ securityRouter.delete('/devices/:deviceId',
     tokensMiddleware,
     async (req: Request, res: Response) => {
     const reqId = req.params.deviceId;
-    if(!reqId) return res.status(404).send({errorsMessage: []})
     const verify = await jwtService.verifyToken(req.cookies.refreshToken);
     const input = await devicesRepository.getSessionById(reqId);
 
-    if(!input) return res.sendStatus(404)
+    if(!input || !reqId) return res.sendStatus(404)
     if (input.userId !== verify.userId) return res.sendStatus(403);
+
     await authRepository.blackList(req.cookies.refreshToken)
     await devicesRepository.deleteSessionById(reqId);
     res.sendStatus(204)
